Move SQL/PostgresConfig/MariaDBConfig from pkg/model into pkg/sqlstore
Per masv3971's review on #571: revisits the earlier "defer to a future
project" decision on where these types should live now that the
supposed blocker doesn't hold. pkg/sqlstore.Connect's *model.SQL
parameter was pkg/sqlstore's only dependency on pkg/model -- moving the
types there entirely removes that import, leaving a clean one-way
pkg/model -> pkg/sqlstore edge instead of the reverse, so no cycle.

- pkg/sqlstore/config.go: SQL, PostgresConfig, MariaDBConfig struct
  definitions, moved verbatim from pkg/model/config.go.
- pkg/sqlstore/dsn.go: DSN()/libpqQuote/tlsConfigName, moved verbatim
  from pkg/model/sql_config.go (deleted).
- pkg/model/config.go: Common.SQL is now sqlstore.SQL.
- pkg/sqlstore/connect.go: Connect takes *SQL (local type) instead of
  *model.SQL; no longer imports pkg/model at all.
- pkg/helpers/validate.go: the two SQL-related struct validations now
  type-assert sqlstore.SQL instead of model.SQL.
- developer_tools/scripts/gen_config_docs: added pkg/sqlstore to the
  scanned source directories (the doc generator resolves types by name
  across a directory list, not by package, so this is the only change
  needed there); regenerated docs/CONFIGURATION.md, confirming the SQL
  config section and both struct-level constraint notes still render
  correctly.

pkg/model/secrets.go needed no changes: it merges SQL secrets via plain
field access (cfg.Common.SQL.Postgres.Password), which works
transparently regardless of which package defines the field's type.
