+++ pkg/sqlstore/dsn.go
@@ -1,4 +1,4 @@
-package model
+package sqlstore
 
 import (
 	"crypto/tls"
